Consolidate under the same API
