+++ pkg/server/router/presentation.go
@@ -6,17 +6,18 @@ import (
 	"github.com/TBD54566975/ssi-sdk/credential/exchange"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
+	"github.com/tbd54566975/ssi-service/internal/keyaccess"
 	"github.com/tbd54566975/ssi-service/pkg/server/framework"
 	svcframework "github.com/tbd54566975/ssi-service/pkg/service/framework"
 	"github.com/tbd54566975/ssi-service/pkg/service/presentation"
 	"net/http"
 )
 
-type PresentationDefinitionRouter struct {
+type PresentationRouter struct {
 	service *presentation.Service
 }
 
-func NewPresentationDefinitionRouter(s svcframework.Service) (*PresentationDefinitionRouter, error) {
+func NewPresentationDefinitionRouter(s svcframework.Service) (*PresentationRouter, error) {
 	if s == nil {
 		return nil, errors.New("service cannot be nil")
 	}
@@ -24,7 +25,7 @@ func NewPresentationDefinitionRouter(s svcframework.Service) (*PresentationDefin
 	if !ok {
 		return nil, fmt.Errorf("could not create presentation router with service type: %s", s.Type())
 	}
-	return &PresentationDefinitionRouter{service: service}, nil
+	return &PresentationRouter{service: service}, nil
 }
 
 type CreatePresentationDefinitionRequest struct {
@@ -50,7 +51,7 @@ type CreatePresentationDefinitionResponse struct {
 // @Failure      400      {string}  string  "Bad request"
 // @Failure      500      {string}  string  "Internal server error"
 // @Router       /v1/presentation/definition [put]
-func (pdr PresentationDefinitionRouter) CreatePresentationDefinition(ctx context.Context, w http.ResponseWriter, r *http.Request) error {
+func (pr PresentationRouter) CreatePresentationDefinition(ctx context.Context, w http.ResponseWriter, r *http.Request) error {
 	var request CreatePresentationDefinitionRequest
 	errMsg := "Invalid Presentation Definition Request"
 	if err := framework.Decode(r, &request); err != nil {
@@ -68,7 +69,7 @@ func (pdr PresentationDefinitionRouter) CreatePresentationDefinition(ctx context
 		logrus.WithError(err).Error(errMsg)
 		return framework.NewRequestError(errors.Wrap(err, errMsg), http.StatusBadRequest)
 	}
-	serviceResp, err := pdr.service.CreatePresentationDefinition(presentation.CreatePresentationDefinitionRequest{PresentationDefinition: *def})
+	serviceResp, err := pr.service.CreatePresentationDefinition(presentation.CreatePresentationDefinitionRequest{PresentationDefinition: *def})
 	if err != nil {
 		logrus.WithError(err).Error(errMsg)
 		return framework.NewRequestError(errors.Wrap(err, errMsg), http.StatusInternalServerError)
@@ -125,7 +126,7 @@ type GetPresentationDefinitionResponse struct {
 // @Success      200  {object}  GetPresentationDefinitionResponse
 // @Failure      400  {string}  string  "Bad request"
 // @Router       /v1/presentation/definition/{id} [get]
-func (pdr PresentationDefinitionRouter) GetPresentationDefinition(ctx context.Context, w http.ResponseWriter, r *http.Request) error {
+func (pr PresentationRouter) GetPresentationDefinition(ctx context.Context, w http.ResponseWriter, r *http.Request) error {
 	id := framework.GetParam(ctx, IDParam)
 	if id == nil {
 		errMsg := "cannot get presentation without ID parameter"
@@ -133,7 +134,7 @@ func (pdr PresentationDefinitionRouter) GetPresentationDefinition(ctx context.Co
 		return framework.NewRequestErrorMsg(errMsg, http.StatusBadRequest)
 	}
 
-	def, err := pdr.service.GetPresentationDefinition(presentation.GetPresentationDefinitionRequest{ID: *id})
+	def, err := pr.service.GetPresentationDefinition(presentation.GetPresentationDefinitionRequest{ID: *id})
 	if err != nil {
 		errMsg := fmt.Sprintf("could not get presentation with id: %s", *id)
 		logrus.WithError(err).Error(errMsg)
@@ -157,7 +158,7 @@ func (pdr PresentationDefinitionRouter) GetPresentationDefinition(ctx context.Co
 // @Failure      400  {string}  string  "Bad request"
 // @Failure      500  {string}  string  "Internal server error"
 // @Router       /v1/presentation/definition/{id} [delete]
-func (pdr PresentationDefinitionRouter) DeletePresentationDefinition(ctx context.Context, w http.ResponseWriter, _ *http.Request) error {
+func (pr PresentationRouter) DeletePresentationDefinition(ctx context.Context, w http.ResponseWriter, _ *http.Request) error {
 	id := framework.GetParam(ctx, IDParam)
 	if id == nil {
 		errMsg := "cannot delete a presentation without an ID parameter"
@@ -165,11 +166,108 @@ func (pdr PresentationDefinitionRouter) DeletePresentationDefinition(ctx context
 		return framework.NewRequestErrorMsg(errMsg, http.StatusBadRequest)
 	}
 
-	if err := pdr.service.DeletePresentationDefinition(presentation.DeletePresentationDefinitionRequest{ID: *id}); err != nil {
+	if err := pr.service.DeletePresentationDefinition(presentation.DeletePresentationDefinitionRequest{ID: *id}); err != nil {
 		errMsg := fmt.Sprintf("could not delete presentation with id: %s", *id)
 		logrus.WithError(err).Error(errMsg)
 		return framework.NewRequestError(errors.Wrap(err, errMsg), http.StatusInternalServerError)
 	}
 
 	return framework.Respond(ctx, w, nil, http.StatusOK)
+}
+
+type CreateSubmissionRequest struct {
+	PresentationJwt keyaccess.JWT `json:"presentationJwt" validate:"required"`
+}
+
+type Operation struct {
+	ID     string          `json:"id"`
+	Done   bool            `json:"bool"`
+	Result OperationResult `json:"result"`
+}
+
+type OperationResult struct {
+	Error    string                          `json:"error"`
+	Response exchange.PresentationSubmission `json:"response"`
+}
+
+// CreateSubmission godoc
+// @Summary      Create Submission
+// @Description  Creates a submission in this server ready to be reviewed.
+// @Tags         SubmissionAPI
+// @Accept       json
+// @Produce      json
+// @Param        request  body      CreateSubmissionRequest  true  "request body"
+// @Success      201      {object}  Operation
+// @Failure      400      {string}  string  "Bad request"
+// @Failure      500      {string}  string  "Internal server error"
+// @Router       /v1/presentations/submissions [put]
+func (pr PresentationRouter) CreateSubmission(ctx context.Context, w http.ResponseWriter, r *http.Request) error {
+	var resp Operation
+	return framework.Respond(ctx, w, resp, http.StatusCreated)
+}
+
+type GetSubmissionResponse struct {
+	Submission exchange.PresentationSubmission `json:"submission"`
+}
+
+// GetSubmission godoc
+// @Summary      Get Submission
+// @Description  Get a submission by its ID
+// @Tags         SubmissionAPI
+// @Accept       json
+// @Produce      json
+// @Param        id   path      string  true  "ID"
+// @Success      200  {object}  GetSubmissionResponse
+// @Failure      400  {string}  string  "Bad request"
+// @Router       /v1/presentations/submission/{id} [get]
+func (pr PresentationRouter) GetSubmission(ctx context.Context, w http.ResponseWriter, r *http.Request) error {
+	var resp GetSubmissionResponse
+	return framework.Respond(ctx, w, resp, http.StatusOK)
+}
+
+type ListSubmissionRequest struct {
+	Filter string `json:"filter"`
+}
+
+type ListSubmissionResponse struct {
+	Submissions []exchange.PresentationSubmission `json:"submissions"`
+}
+
+// ListSubmissions godoc
+// @Summary      List Submissions
+// @Description  List existing submissions according to a filtering query.
+// @Tags         SubmissionAPI
+// @Accept       json
+// @Produce      json
+// @Param        request  body      ListSubmissionRequest  true  "request body"
+// @Success      200  {object}  ListSubmissionResponse
+// @Failure      400  {string}  string  "Bad request"
+// @Failure      500  {string}  string  "Internal server error"
+// @Router       /v1/presentations/submissions [get]
+func (pr PresentationRouter) ListSubmissions(ctx context.Context, w http.ResponseWriter, r *http.Request) error {
+	return framework.Respond(ctx, w, ListSubmissionRequest{}, http.StatusOK)
+}
+
+type ReviewSubmissionRequest struct {
+	Approved bool   `json:"approved" validate:"required"`
+	Reason   string `json:"reason"`
+}
+
+type ReviewSubmissionResponse struct {
+	Submission exchange.PresentationSubmission `json:"submission"`
+}
+
+// ReviewSubmission godoc
+// @Summary      Review a pending submissions
+// @Description  Reviews a pending submission. After this method is called, the operation with `id==presentations/submissions/{submission_id}` will be updated with the result of this invocation.
+// @Tags         SubmissionAPI
+// @Accept       json
+// @Produce      json
+// @Param        request  body      ReviewSubmissionRequest  true  "request body"
+// @Success      200  {object}  ReviewSubmissionResponse
+// @Failure      400  {string}  string  "Bad request"
+// @Failure      500  {string}  string  "Internal server error"
+// @Router       /v1/presentations/submissions [get]
+func (pr PresentationRouter) ReviewSubmission(ctx context.Context, w http.ResponseWriter, r *http.Request) error {
+	return framework.Respond(ctx, w, ReviewSubmissionResponse{}, http.StatusOK)
 }
