+++ pkg/server/server_presentation_test.go
@@ -40,7 +40,7 @@ func TestPresentationAPI(t *testing.T) {
 	s, err := storage.NewStorage(storage.Bolt)
 	assert.NoError(t, err)
 
-	service, err := presentation.NewPresentationDefinitionService(config.PresentationServiceConfig{}, s)
+	service, err := presentation.NewPresentationService(config.PresentationServiceConfig{}, s)
 	assert.NoError(t, err)
 
 	pRouter, err := router.NewPresentationDefinitionRouter(service)
