+++ pkg/service/framework/framework.go
@@ -14,7 +14,6 @@ const (
 	KeyStore     Type = "keystore"
 	Manifest     Type = "manifest"
 	Presentation Type = "presentation"
-	Submission   Type = "submission"
 
 	StatusReady    StatusState = "ready"
 	StatusNotReady StatusState = "not_ready"
