+++ pkg/service/presentation/model.go
@@ -29,3 +29,27 @@ type GetPresentationDefinitionResponse struct {
 type DeletePresentationDefinitionRequest struct {
 	ID string `json:"id" validate:"required"`
 }
+
+type CreateSubmissionRequest struct {
+	Submission exchange.PresentationSubmission `json:"submission" validate:"required"`
+}
+
+func (csr CreateSubmissionRequest) IsValid() bool {
+	return util.IsValidStruct(csr) == nil
+}
+
+type CreateSubmissionResponse struct {
+	Submission exchange.PresentationSubmission `json:"submission"`
+}
+
+type GetSubmissionRequest struct {
+	ID string `json:"id" validate:"required"`
+}
+
+type GetSubmissionResponse struct {
+	Submission exchange.PresentationSubmission `json:"submission"`
+}
+
+type DeleteSubmissionRequest struct {
+	ID string `json:"id" validate:"required"`
+}
