+++ pkg/service/presentation/storage/bolt.go
@@ -11,7 +11,8 @@ import (
 )
 
 const (
-	namespace = "presentation_definition"
+	namespace           = "presentation_definition"
+	submissionNamespace = "presentation_submission"
 )
 
 type BoltPresentationStorage struct {
@@ -69,4 +70,41 @@ func (b BoltPresentationStorage) DeletePresentation(id string) error {
 		return errors.Wrapf(err, errMsg)
 	}
 	return nil
+}
+
+func (b BoltPresentationStorage) StoreSubmission(submission StoredSubmission) error {
+	id := submission.Submission.ID
+	if id == "" {
+		err := errors.New("could not store submission definition without an ID")
+		logrus.WithError(err).Error()
+		return err
+	}
+	jsonBytes, err := json.Marshal(submission)
+	if err != nil {
+		errMsg := fmt.Sprintf("could not store submission definition: %s", id)
+		logrus.WithError(err).Error(errMsg)
+		return errors.Wrapf(err, errMsg)
+	}
+	return b.db.Write(submissionNamespace, id, jsonBytes)
+}
+
+func (b BoltPresentationStorage) GetSubmission(id string) (*StoredSubmission, error) {
+	jsonBytes, err := b.db.Read(submissionNamespace, id)
+	if err != nil {
+		errMsg := fmt.Sprintf("could not get submission definition: %s", id)
+		logrus.WithError(err).Error(errMsg)
+		return nil, errors.Wrapf(err, errMsg)
+	}
+	if len(jsonBytes) == 0 {
+		err := fmt.Errorf("submission definition not found with id: %s", id)
+		logrus.WithError(err).Error("could not get submission definition from storage")
+		return nil, err
+	}
+	var stored StoredSubmission
+	if err := json.Unmarshal(jsonBytes, &stored); err != nil {
+		errMsg := fmt.Sprintf("could not unmarshal stored submission definition: %s", id)
+		logrus.WithError(err).Error(errMsg)
+		return nil, errors.Wrapf(err, errMsg)
+	}
+	return &stored, nil
 }
