+++ pkg/service/presentation/storage/storage.go
@@ -12,6 +12,11 @@ type StoredPresentation struct {
 }
 
 type Storage interface {
+	DefinitionStorage
+	SubmissionStorage
+}
+
+type DefinitionStorage interface {
 	StorePresentation(schema StoredPresentation) error
 	GetPresentation(id string) (*StoredPresentation, error)
 	DeletePresentation(id string) error
@@ -33,4 +38,13 @@ func NewPresentationStorage(s storage.ServiceStorage) (Storage, error) {
 	default:
 		return nil, util.LoggingNewErrorf("unsupported storage type: %s", s.Type())
 	}
+}
+
+type StoredSubmission struct {
+	Submission exchange.PresentationSubmission `json:"submission"`
+}
+
+type SubmissionStorage interface {
+	StoreSubmission(schema StoredSubmission) error
+	GetSubmission(id string) (*StoredSubmission, error)
 }
