+++ pkg/service/service.go
@@ -2,9 +2,6 @@ package service
 
 import (
 	"fmt"
-	"github.com/tbd54566975/ssi-service/pkg/service/presentation"
-	"github.com/tbd54566975/ssi-service/pkg/service/submission"
-
 	"github.com/tbd54566975/ssi-service/config"
 	"github.com/tbd54566975/ssi-service/internal/util"
 	"github.com/tbd54566975/ssi-service/pkg/service/credential"
@@ -12,6 +9,7 @@ import (
 	"github.com/tbd54566975/ssi-service/pkg/service/framework"
 	"github.com/tbd54566975/ssi-service/pkg/service/keystore"
 	"github.com/tbd54566975/ssi-service/pkg/service/manifest"
+	"github.com/tbd54566975/ssi-service/pkg/service/presentation"
 	"github.com/tbd54566975/ssi-service/pkg/service/schema"
 	"github.com/tbd54566975/ssi-service/pkg/storage"
 )
@@ -56,9 +54,6 @@ func validateServiceConfig(config config.ServicesConfig) error {
 	if config.PresentationConfig.IsEmpty() {
 		return fmt.Errorf("%s no config provided", framework.Presentation)
 	}
-	if config.SubmissionConfig.IsEmpty() {
-		return fmt.Errorf("%s no config provided", framework.Submission)
-	}
 	return nil
 }
 
@@ -100,15 +95,10 @@ func instantiateServices(config config.ServicesConfig) ([]framework.Service, err
 		return nil, util.LoggingErrorMsg(err, "could not instantiate the manifest service")
 	}
 
-	presentationService, err := presentation.NewPresentationDefinitionService(config.PresentationConfig, storageProvider)
+	presentationService, err := presentation.NewPresentationService(config.PresentationConfig, storageProvider)
 	if err != nil {
 		return nil, util.LoggingErrorMsg(err, "could not instantiate the presentation service")
 	}
 
-	submissionService, err := submission.NewSubmissionService(config.SubmissionConfig, storageProvider)
-	if err != nil {
-		return nil, util.LoggingErrorMsg(err, "could not instantiate the presentation service")
-	}
-
-	return []framework.Service{keyStoreService, didService, schemaService, credentialService, manifestService, presentationService, submissionService}, nil
+	return []framework.Service{keyStoreService, didService, schemaService, credentialService, manifestService, presentationService}, nil
 }
