--- pkg/service/submission/service.go
@@ -1,98 +1,0 @@
-package submission
-
-import (
-	"fmt"
-	"github.com/TBD54566975/ssi-sdk/credential/exchange"
-	sdkutil "github.com/TBD54566975/ssi-sdk/util"
-	"github.com/pkg/errors"
-	"github.com/sirupsen/logrus"
-
-	"github.com/tbd54566975/ssi-service/config"
-	"github.com/tbd54566975/ssi-service/internal/util"
-	"github.com/tbd54566975/ssi-service/pkg/service/framework"
-	submissionstorage "github.com/tbd54566975/ssi-service/pkg/service/submission/storage"
-	"github.com/tbd54566975/ssi-service/pkg/storage"
-)
-
-type Service struct {
-	storage submissionstorage.Storage
-	config  config.SubmissionServiceConfig
-}
-
-func (s Service) Type() framework.Type {
-	return framework.Submission
-}
-
-func (s Service) Status() framework.Status {
-	ae := sdkutil.NewAppendError()
-	if s.storage == nil {
-		ae.AppendString("no storage configured")
-	}
-	if !ae.IsEmpty() {
-		return framework.Status{
-			Status:  framework.StatusNotReady,
-			Message: fmt.Sprintf("submission service is not ready: %s", ae.Error().Error()),
-		}
-	}
-	return framework.Status{Status: framework.StatusReady}
-}
-
-func (s Service) Config() config.SubmissionServiceConfig {
-	return s.config
-}
-
-func NewSubmissionService(config config.SubmissionServiceConfig, s storage.ServiceStorage) (*Service, error) {
-	submissionStorage, err := submissionstorage.NewSubmissionStorage(s)
-	if err != nil {
-		errMsg := "could not instantiate storage for the submission service"
-		return nil, util.LoggingErrorMsg(err, errMsg)
-	}
-	service := Service{
-		storage: submissionStorage,
-		config:  config,
-	}
-	if !service.Status().IsReady() {
-		return nil, errors.New(service.Status().Message)
-	}
-	return &service, nil
-}
-
-// CreateSubmission houses the main service logic for presentation submission creation. It validates the input, and
-// produces a presentation submission value that conforms with the Submission specification.
-func (s Service) CreateSubmission(request CreateSubmissionRequest) (*CreateSubmissionResponse, error) {
-	logrus.Debugf("creating presentation submission: %+v", request)
-
-	if !request.IsValid() {
-		errMsg := fmt.Sprintf("invalid create presentation submission request: %+v", request)
-		return nil, util.LoggingNewError(errMsg)
-	}
-
-	if err := exchange.IsValidPresentationSubmission(request.Submission); err != nil {
-		return nil, util.LoggingErrorMsg(err, "provided value is not a valid presentation submission")
-	}
-
-	storedSubmission := submissionstorage.StoredSubmission{Submission: request.Submission}
-
-	if err := s.storage.StoreSubmission(storedSubmission); err != nil {
-		return nil, util.LoggingErrorMsg(err, "could not store presentation")
-	}
-
-	return &CreateSubmissionResponse{
-		Submission: storedSubmission.Submission,
-	}, nil
-}
-
-func (s Service) GetSubmission(request GetSubmissionRequest) (*GetSubmissionResponse, error) {
-	logrus.Debugf("getting presentation submission: %s", request.ID)
-
-	storedSubmission, err := s.storage.GetSubmission(request.ID)
-	if err != nil {
-		err := errors.Wrapf(err, "error getting presentation submission: %s", request.ID)
-		return nil, util.LoggingError(err)
-	}
-	if storedSubmission == nil {
-		err := fmt.Errorf("presentation submission with id<%s> could not be found", request.ID)
-		return nil, util.LoggingError(err)
-	}
-	return &GetSubmissionResponse{Submission: storedSubmission.Submission}, nil
-}
