--- pkg/service/submission/storage/storage.go
@@ -1,37 +1,0 @@
-package storage
-
-import (
-	"fmt"
-	"github.com/TBD54566975/ssi-sdk/credential/exchange"
-	"github.com/tbd54566975/ssi-service/internal/util"
-	"github.com/tbd54566975/ssi-service/pkg/storage"
-)
-
-type StoredSubmission struct {
-	Submission exchange.PresentationSubmission `json:"submission"`
-}
-
-type Storage interface {
-	StoreSubmission(schema StoredSubmission) error
-	GetSubmission(id string) (*StoredSubmission, error)
-}
-
-// NewSubmissionStorage finds the submission storage impl for a given ServiceStorage value
-func NewSubmissionStorage(s storage.ServiceStorage) (Storage, error) {
-	switch s.Type() {
-	case storage.Bolt:
-		gotBolt, ok := s.(*storage.BoltDB)
-		if !ok {
-			errMsg := fmt.Sprintf("trouble instantiating : %s", s.Type())
-			return nil, util.LoggingNewError(errMsg)
-		}
-		boltStorage, err := NewBoltSubmissionStorage(gotBolt)
-		if err != nil {
-			return nil, util.LoggingErrorMsg(err, "could not instantiate schema bolt storage")
-		}
-		return boltStorage, err
-	default:
-		errMsg := fmt.Errorf("unsupported storage type: %s", s.Type())
-		return nil, util.LoggingError(errMsg)
-	}
-}
